app/requirements.txt
fastapi
uvicorn
cohere
langchain
faiss-cpu
python-dotenv
tiktoken
PyPDF2
